feat: extend atom/select customisation
