+++ resources/assets/js/components/atom/Select/index.js
@@ -3,6 +3,20 @@ import InputLabel from '~/components/atom/InputLabel';
 import Option from '~/components/atom/Option/index';
 import SelectElement from '@material-ui/core/Select';
 import FormControl, { useStyles } from './style';
+import isString from 'lodash/isString';
+import isArray from 'lodash/isArray';
+
+function combineClassName(rootClass, propsClass) {
+    if (propsClass !== undefined) {
+        if (isString(propsClass)) {
+            return `${rootClass} ${propsClass}`;
+        } else if (isArray(propsClass)) {
+            return `${rootClass} ${propsClass.join(' ')}`;
+        }
+    } else {
+        return rootClass;
+    }
+}
 
 function Select(props) {
     const classes = useStyles();
@@ -14,8 +28,9 @@ function Select(props) {
                 value={props.value}
                 {...props.elementAttrs}
                 onChange={props.onChange}
+                MenuProps={props.menuProps}
                 SelectDisplayProps={{ className: classes.selectDisplay }}
-                className={classes.root}
+                className={combineClassName(classes.root, props.classes)}
             >
                 {props.options &&
                     props.options.map((item) => (
